+++ R/batch.R
@@ -152,24 +152,19 @@ isvacor <- function(data, lv) {
         return(li)
 }
 
-#' Principal component analysis(PCA) for SVA corrected data and raw data
+#' Principal component analysis(PCA) for SVA/ISVA corrected data and raw data
 #' @param list results from svacor function
 #' @param center parameters for PCA
 #' @param scale parameters for scale
-#' @param lv group information
+#' @param lv factor vector for the group infomation
 #' @return plot
 #' @examples
 #' \dontrun{
 #' library(faahKO)
 #' cdfpath <- system.file("cdf", package = "faahKO")
-#' cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
-#' xset <- xcmsSet(cdffiles)
-#' xset <- group(xset)
-#' xset2 <- retcor(xset, family = "symmetric", plottype = "mdevden")
-#' xset2 <- group(xset2, bw = 10)
-#' xset3 <- fillPeaks(xset2)
-#' df <- svacor(xset3)
-#' svapca(df)
+#' list <- getmr(cdfpath, pmethod = ' ')
+#' li <- svacor(list$data,list$group$class)
+#' svapca(li)
 #' }
 #' @seealso \code{\link{svacor}}, \code{\link{svaplot}}, \code{\link{svabatch}}
 #' @export
@@ -182,7 +177,7 @@ svapca <- function(list, center = T, scale = T, lv = NULL) {
     if (is.null(lv)) {
         pch = colnames(data)
     } else {
-        pch = lv
+        pch = as.character(lv)
     }
 
     graphics::par(mfrow = c(2, 5), mar = c(4, 4, 2.6,
@@ -243,140 +238,51 @@ svapca <- function(list, center = T, scale = T, lv = NULL) {
         pcacVars[2], "% of Variance Explained"), pch = pch,
         cex = 2, main = "PCA-corrected")
 }
-#' Filter the data with p value and q value
-#' @param list results from svacor function
-#' @param pqvalues method for ANOVA or SVA
-#' @param pt threshold for p value, default is 0.05
-#' @param qt threshold for q value, default is 0.05
-#' @return data, corrected data, mz and retention for fileted data
-#' @examples
-#' \dontrun{
-#' library(faahKO)
-#' cdfpath <- system.file("cdf", package = "faahKO")
-#' cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
-#' xset <- xcmsSet(cdffiles)
-#' xset <- group(xset)
-#' xset2 <- retcor(xset, family = "symmetric", plottype = "mdevden")
-#' xset2 <- group(xset2, bw = 10)
-#' xset3 <- fillPeaks(xset2)
-#' df <- svacor(xset3)
-#' svadata(df)
-#' }
-#' @export
-svadata <- function(list, pqvalues = "sv", pt = 0.05,
-    qt = 0.05) {
-    data <- list$data
-    signal2 <- list$signal2
-    datacor <- list$dataCorrected
-    pValues <- list$"p-values"
-    qValues <- list$"q-values"
-    pValuesSv <- list$"p-valuesCorrected"
-    qValuesSv <- list$"q-valuesCorrected"
-    mz <- list$mz
-    rt <- list$rt
-    if (is.null(signal2)) {
-        if (pqvalues == "anova" & sum(pValues < pt &
-            qValues < qt) != 0) {
-            message("No SV while p-values and q-values have results")
-            data <- data[pValues < pt & qValues < qt,
-                ]
-            mz <- mz[pValues < pt & qValues < qt]
-            rt <- rt[pValues < pt & qValues < qt]
-            li <- list(data, pValues < pt & qValues <
-                qt, mz, rt)
-            names(li) <- c("data", "pqvalues", "mz",
-                "rt")
-            return(li)
-        } else {
-            message("No SV while p-values and q-values have no results")
-        }
-    } else {
-        if (pqvalues == "anova" & sum(pValues < pt &
-            qValues < qt) != 0) {
-            message("Have SVs while p-values and q-values have results")
-            data <- data[pValues < pt & qValues < qt,
-                ]
-            mz <- mz[pValues < pt & qValues < qt]
-            rt <- rt[pValues < pt & qValues < qt]
-            li <- list(data, pValues < pt & qValues <
-                qt, mz, rt)
-            names(li) <- c("data", "pqvalues", mz,
-                rt)
-            return(li)
-        } else if (pqvalues == "anova") {
-            message("Have SVs while p-values and q-values have no results")
-        } else if (pqvalues == "sv" & sum(pValuesSv <
-            pt & qValuesSv < qt) != 0) {
-            message("SVs corrected while p-values and q-values have results")
-            data <- data[pValuesSv < pt & qValuesSv <
-                qt, ]
-            datacor <- datacor[pValuesSv < pt & qValuesSv <
-                qt, ]
-            mz <- mz[pValuesSv < pt & qValuesSv < qt]
-            rt <- rt[pValuesSv < pt & qValuesSv < qt]
-            li <- list(datacor, data, pValuesSv < pt &
-                qValuesSv < qt, mz, rt)
-            names(li) <- c("dataCorrected", "data",
-                "pqvalues", "mz", "rt")
-            return(li)
-        } else {
-            message("SVs corrected while p-values and q-values have no results")
-        }
-    }
-}
+
 #' Filter the data with p value and q value and show them
 #' @param list results from svacor function
+#' @param lv factor vector for the group infomation
 #' @param pqvalues method for ANOVA or SVA
 #' @param pt threshold for p value, default is 0.05
 #' @param qt threshold for q value, default is 0.05
-#' @param lv group information
 #' @param index index for selected peaks
 #' @return heatmap for the data
 #' @examples
 #' \dontrun{
 #' library(faahKO)
 #' cdfpath <- system.file("cdf", package = "faahKO")
-#' cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
-#' xset <- xcmsSet(cdffiles)
-#' xset <- group(xset)
-#' xset2 <- retcor(xset, family = "symmetric", plottype = "mdevden")
-#' xset2 <- group(xset2, bw = 10)
-#' xset3 <- fillPeaks(xset2)
-#' df <- svacor(xset3)
-#' svaplot(df)
+#' list <- getmr(cdfpath, pmethod = ' ')
+#' li <- svacor(list$data,list$group$class)
+#' svaplot(li,list$group$class)
 #' }
 #' @seealso \code{\link{svacor}}, \code{\link{svapca}}, \code{\link{svabatch}}
 #' @export
-svaplot <- function(list, pqvalues = "sv", pt = 0.05,
-    qt = 0.05, lv = NULL, index = NULL) {
-        if (is.null(lv)) {
-                lv <- colnames(data)
-        }
-    data <- list$data
-    signal <- list$signal
-    signal2 <- list$signal2
-    batch <- list$batch
-    error <- list$error
-    error2 <- list$error2
-    datacor <- list$dataCorrected
+svaplot <- function(list, lv, pqvalues = "sv", pt = 0.05,
+    qt = 0.05, index = NULL) {
+    data <- list$data[,order(lv)]
+    signal <- list$signal[,order(lv)]
+    signal2 <- list$signal2[,order(lv)]
+    batch <- list$batch[,order(lv)]
+    error <- list$error[,order(lv)]
+    error2 <- list$error2[,order(lv)]
+    datacor <- list$dataCorrected[,order(lv)]
     pValues <- list$"p-values"
     qValues <- list$"q-values"
     pValuesSv <- list$"p-valuesCorrected"
     qValuesSv <- list$"q-valuesCorrected"
     if (!is.null(index)) {
-        data <- data[index,]
-        signal <- signal[index,]
-        signal2 <- signal2[index,]
-        batch <- batch[index,]
-        error <- error[index,]
-        error2 <- error2[index,]
-        datacor <- datacor[index,]
+        data <- data[index,order(lv)]
+        signal <- signal[index,order(lv)]
+        signal2 <- signal2[index,order(lv)]
+        batch <- batch[index,order(lv)]
+        error <- error[index,order(lv)]
+        error2 <- error2[index,order(lv)]
+        datacor <- datacor[index,order(lv)]
         pValues <- pValues[index]
         qValues <- qValues[index]
         pValuesSv <- pValuesSv[index]
         qValuesSv <- qValuesSv[index]
     }
-    lv <- as.factor(lv)
     pos <- cumsum(as.numeric(table(lv)/sum(table(lv)))) -
         as.numeric(table(lv)/sum(table(lv)))/2
     posv <- cumsum(as.numeric(table(lv)/sum(table(lv))))[1:(nlevels(lv) -1)]
@@ -590,41 +496,6 @@ svaplot <- function(list, pqvalues = "sv", pt = 0.05,
     }
 }
 
-#' Get the corrected data after SVA for metabolanalyst
-#' @param xset xcmsset object
-#' @param lv group information
-#' @return csv files for both raw and corrected data for metabolanalyst if SVA could be applied
-#' @examples
-#' \dontrun{
-#' library(faahKO)
-#' cdfpath <- system.file("cdf", package = "faahKO")
-#' cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
-#' xset <- xcmsSet(cdffiles)
-#' xset <- group(xset)
-#' xset2 <- retcor(xset, family = "symmetric", plottype = "mdevden")
-#' xset2 <- group(xset2, bw = 10)
-#' xset3 <- fillPeaks(xset2)
-#' svaupload(xset3)
-#' }
-#' @export
-svaupload <- function(xset, lv = NULL) {
-    raw <- svacor(xset, lv = lv)
-    if (is.null(raw$dataCorrected)) {
-        data <- raw$data
-        data <- rbind(group = as.character(xset@phenoData[,
-            1]), data)
-        utils::write.csv(data, file = "Peaklist.csv")
-        return(data)
-    } else {
-        datacor <- raw$dataCorrected
-        data <- raw$data
-        data <- rbind(group = as.character(xset@phenoData[,
-            1]), data)
-        utils::write.csv(datacor, file = "Peaklistcor.csv")
-        utils::write.csv(data, file = "Peaklist.csv")
-        return(list(data, datacor))
-    }
-}
 
 #' Plot the influnces of DoE and Batch effects on each peaks
 #' @param df data output from `svacor` function
@@ -635,16 +506,11 @@ svaupload <- function(xset, lv = NULL) {
 #' \dontrun{
 #' library(faahKO)
 #' cdfpath <- system.file("cdf", package = "faahKO")
-#' cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
-#' xset <- xcmsSet(cdffiles)
-#' xset <- group(xset)
-#' xset2 <- retcor(xset, family = "symmetric", plottype = "mdevden")
-#' xset2 <- group(xset2, bw = 10)
-#' xset3 <- fillPeaks(xset2)
-#' df <- svacor(xset3)
-#' dfsv <- svaplot(xset3)
-#' dfanova <- svaplot(xset3, pqvalues = "anova")
-#' svabatch(df,dfsv,dfanova)
+#' list <- getmr(cdfpath, pmethod = ' ')
+#' li <- svacor(list$data,list$group$class)
+#' dfsv <- svaplot(li,list$group$class)
+#' dfanova <- svaplot(li, list$group$class, pqvalues = "anova")
+#' svabatch(li,dfsv,dfanova)
 #' }
 #' @seealso \code{\link{svacor}}, \code{\link{svaplot}}, \code{\link{svapca}}
 #' @export